pyproject.toml
[project]
name = "rprag"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "datasets>=4.1.1",
    "hydra-core>=1.3.2",
    "omegaconf>=2.3.0",
    "torch>=2.8.0",
    "vllm>=0.10.2",
]
